requirements.txt
requests==2.25.1
gate_api==4.22.2
PyYAML==5.4.1
pytest==4.6.11